dynamodb

+++ src/test/java/de/wirvsvirus/hack/dynamodb/DynamoDBIntegrationTest.java
@@ -24,6 +24,11 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+/**
+ * https://docs.aws.amazon.com/de_de/amazondynamodb/latest/developerguide/CodeSamples.Java.html
+ * https://docs.aws.amazon.com/de_de/amazondynamodb/latest/developerguide/DynamoDBMapper.CRUDExample1.html
+ *
+ */
 @ExtendWith (SpringExtension.class)
 @SpringBootTest(classes = DynamoDBConfiguration.class)
 @TestPropertySource(properties = {
@@ -41,26 +46,27 @@ public class DynamoDBIntegrationTest {
     public void setup() throws Exception {
         dynamoDBMapper = new DynamoDBMapper(amazonDynamoDB);
 
-        DeleteTableRequest deleteTableRequest = dynamoDBMapper
-                .generateDeleteTableRequest(ProductInfo.class);
+        if (false) {
+            DeleteTableRequest deleteTableRequest = dynamoDBMapper
+                    .generateDeleteTableRequest(ProductInfo.class);
 
-        try {
-            amazonDynamoDB.deleteTable(deleteTableRequest);
-        } catch(ResourceNotFoundException rnfe) {
+            try {
+                amazonDynamoDB.deleteTable(deleteTableRequest);
+            } catch(ResourceNotFoundException rnfe) {
+            }
+            CreateTableRequest tableRequest = dynamoDBMapper
+                    .generateCreateTableRequest(ProductInfo.class);
+
+            tableRequest.setProvisionedThroughput(
+                    new ProvisionedThroughput(1L, 1L));
+            amazonDynamoDB.createTable(tableRequest);
         }
-
-        CreateTableRequest tableRequest = dynamoDBMapper
-                .generateCreateTableRequest(ProductInfo.class);
-
-        tableRequest.setProvisionedThroughput(
-                new ProvisionedThroughput(1L, 1L));
-        amazonDynamoDB.createTable(tableRequest);
 
         for (int i=10; i<99;i++) {
             ProductInfo dave = new ProductInfo();
             dave.setCost("25");
             dave.setMsrp("zzzz " + i);
-            dynamoDBMapper.save(dave);
+//            dynamoDBMapper.save(dave);
         }
 
         //...
@@ -78,6 +84,18 @@ public class DynamoDBIntegrationTest {
     }
 
     @Test
+    void getItem() {
+
+        final ProductInfo p1224 = new ProductInfo();
+        p1224.setId("1224");
+        p1224.setCost("32");
+        p1224.setMsrp("222");
+        dynamoDBMapper.save(p1224);
+        final ProductInfo productInfo = dynamoDBMapper.load(ProductInfo.class, "324d4db2-edf1-4292-b1ef-27c59db66f25");
+        System.out.println("> " + productInfo);
+    }
+
+    @Test
     void findall() {
         Map<String, AttributeValue> eav = new HashMap<String, AttributeValue>();
         eav.put(":val1", new AttributeValue().withS("e53abae7-69e7-42b4-bdee-6d392b2ed80d"));
@@ -89,6 +107,6 @@ public class DynamoDBIntegrationTest {
                 ;
 
         final PaginatedScanList<ProductInfo> result = dynamoDBMapper.scan(ProductInfo.class, scan);
-        result.forEach(p -> System.out.println("- " + p.getMsrp()));
+        result.forEach(p -> System.out.println("- " + p.getId()));
     }
 }
